+++ main.cpp
@@ -1,6 +1,7 @@
 #include "romreader.h"
 #include "cpu.h"
 #include <iostream>
+#include <utility> 
 #include <fstream>
 #include <string>
 
@@ -9,18 +10,17 @@
 //
 
 int main() {
-    
 	Cpu cpu_;
-    Romreader romreader_;
+	Romreader romreader_;
 
-    romreader_.read_file("smb.nes");
+	romreader_.read_file("smb.nes");
 	for (int i = 0; i < 100; i++) {
-		std::pair<std::uint_least8_t, std::uint_least8_t> opcode = romreader_.return_opcode(i);
-	    cpu_.disassemble(opcode);
+		std::pair<std::uint_least8_t, std::uint_least16_t> opcode = romreader_.return_opcode(i);
+		cpu_.disassemble(opcode);
 	}
 
 	std::getchar();
     
-    return 0;
+	return 0;
 }
